fixing types

+++ DataFormats/interface/ElectronCandidateFeatures.h
@@ -67,7 +67,7 @@ class ElectronCandidateFeatures {
     float elec_dz;
     int elec_nbOfMissingHits;
     float elec_gsfCharge;
-    float elec_ndof ; 
+    int   elec_ndof ; 
     float elec_chi2 ;
